requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2>=2.8.6
python-multipart>=0.0.6
pydantic>=2.4.2
pydantic-settings>=2.0.0
python-jose>=3.3.0
python-dotenv>=1.0.0
qrcode>=7.4.2
Pillow>=10.0.0
passlib>=1.7.4
bcrypt>=4.0.1
email-validator>=2.0.0
